+++ apis/auth-api/src/config/db.config.ts
@@ -1,7 +1,6 @@
 import { registerAs } from "@nestjs/config";
-import { req } from "./env.util";
 
 export const dbConfig = registerAs("db", () => ({
-	url: req("DATABASE_URL"),
+	url: process.env.DATABASE_URL,
 }));
 export type DbConfig = ReturnType<typeof dbConfig>;