+++ apps/account-server/src/lib/account.ts
@@ -13,6 +13,9 @@ import { env } from '@/env';
 import { CONTRACTS, SOPHON_VIEM_CHAIN } from './constants';
 import { SOPHON_SALT_PREFIX } from './smart-contract';
 
+const MAX_RETRIES = 5;
+const RETRY_DELAY = 500;
+
 // this part should be on the backend
 export const deployAccount = async (ownerAddress: `0x${string}`) => {
   const deployerAccount = privateKeyToAccount(
@@ -26,15 +29,26 @@ export const deployAccount = async (ownerAddress: `0x${string}`) => {
       transport: http(),
     }).extend(eip712WalletActions());
 
-  const deployedAccount = await deployModularAccount(deployerClient, {
-    accountFactory: CONTRACTS.accountFactory,
-    paymaster: {
-      location: CONTRACTS.accountPaymaster,
-    },
-    uniqueAccountId: SOPHON_SALT_PREFIX,
-    owners: [ownerAddress],
-    installNoDataModules: [],
-  });
+  let retries = MAX_RETRIES;
 
-  return deployedAccount;
+  do {
+    try {
+      const deployedAccount = await deployModularAccount(deployerClient, {
+        accountFactory: CONTRACTS.accountFactory,
+        paymaster: {
+          location: CONTRACTS.accountPaymaster,
+        },
+        uniqueAccountId: SOPHON_SALT_PREFIX,
+        owners: [ownerAddress],
+        installNoDataModules: [],
+      });
+
+      return deployedAccount;
+    } catch (error) {
+      console.error(error);
+      await new Promise((resolve) => setTimeout(resolve, RETRY_DELAY));
+    }
+  } while (--retries > 0);
+
+  throw new Error(`Failed to deploy account for owner: ${ownerAddress}`);
 };