+++ apps/account-server/src/providers/Web3Provider.tsx
@@ -1,23 +1,12 @@
 'use client';
 import { EthereumWalletConnectors } from '@dynamic-labs/ethereum';
-import {
-  DynamicContextProvider,
-  dynamicEvents,
-} from '@dynamic-labs/sdk-react-core';
+import { DynamicContextProvider } from '@dynamic-labs/sdk-react-core';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { type ReactNode, useState } from 'react';
 import { type State, WagmiProvider } from 'wagmi';
 import { env } from '@/env';
 import { sendMessage } from '@/events';
 import { getWagmiConfig } from './wagmi';
-
-// work around for dynamic events not working properly and giving the primary wallet on auth success
-dynamicEvents.on('primaryWalletChanged', (newPrimaryWallet) => {
-  sendMessage('k1.login', {
-    address: newPrimaryWallet.address as `0x${string}`,
-    wallet: newPrimaryWallet,
-  });
-});
 
 export const Web3Provider = ({
   children,
@@ -54,7 +43,13 @@ export const Web3Provider = ({
             sendMessage('k1.login.init', null);
           },
           // This is called after everything is validated and the user should be considered authenticated
-          onAuthSuccess() {
+          onAuthSuccess({ primaryWallet }) {
+            if (primaryWallet) {
+              sendMessage('k1.login', {
+                address: primaryWallet.address as `0x${string}`,
+                wallet: primaryWallet,
+              });
+            }
             sendMessage('modal.open', null);
           },
           // This is called when the user