+++ apps/account-server/src/state/state-machine.ts
@@ -193,6 +193,9 @@ export const userWalletRequestStateMachine = createMachine({
             ACCOUNT_ERROR: {
               target: 'idle',
             },
+            LOGIN_SUCCESS: {
+              target: 'complete',
+            },
           },
         },
         deployment: {
@@ -359,22 +362,15 @@ export const userWalletRequestStateMachine = createMachine({
       on: {
         ACCEPT: {
           target: 'completed',
-          actions: 'clearRequests',
+          actions: 'logout',
         },
         CANCEL: {
           target: 'completed',
-          actions: 'cancelRequests',
+          actions: 'logout',
         },
         LOGOUT: {
           target: 'login-required',
-          actions: [
-            'clearRequests',
-            assign(({ context }) => ({
-              ...context,
-              isAuthenticated: false,
-              isLoadingResources: false,
-            })),
-          ],
+          actions: 'logout',
         },
       },
     },