+++ examples/web/eip6963/components/Web3Provider.tsx
@@ -8,10 +8,7 @@ import { sophon, sophonTestnet } from 'viem/chains';
 import { createConfig, http, type State, WagmiProvider } from 'wagmi';
 
 const CHAIN_ID = (process.env.NEXT_PUBLIC_CHAIN_ID as string) ?? '531050104';
-createSophonEIP6963Emitter(
-  CHAIN_ID === '50104' ? 'mainnet' : 'testnet',
-  //'http://localhost:3000',
-);
+createSophonEIP6963Emitter(CHAIN_ID === '50104' ? 'mainnet' : 'testnet');
 
 export const projectId = '760fb7a448e58431c9cfbab80743ab1c';