+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-eip6963
 
+## 2.1.2
+
+### Patch Changes
+
+- Bugfixes and disconnect/logout support
+- Updated dependencies
+  - @sophon-labs/account-provider@2.1.2
+  - @sophon-labs/account-core@2.1.2
+
 ## 2.1.1
 
 ### Patch Changes